move clustering helper functions to cluster_util and rmeove all global variables
